Sample SE.MATS Skipped-Exon Splicing Events

A dataset of skipped exon (SE) alternative splicing events generated by rMATS, provided for testing and demonstration of the splicing map and sequence motif analysis functions in RNAPeaks.

Usage

sample_se.mats

Format

A data frame with 87,736 observations and 23 variables:

ID

Event ID

GeneID

Ensembl gene ID

geneSymbol

Gene symbol

chr

Chromosome

strand

Strand (+ or -)

exonStart_0base

Skipped exon start position (0-based)

exonEnd

Skipped exon end position

upstreamES

Upstream exon start position

upstreamEE

Upstream exon end position

downstreamES

Downstream exon start position

downstreamEE

Downstream exon end position

ID.1

Duplicate ID column

IJC_SAMPLE_1

Inclusion junction counts for sample 1

SJC_SAMPLE_1

Skipping junction counts for sample 1

IJC_SAMPLE_2

Inclusion junction counts for sample 2

SJC_SAMPLE_2

Skipping junction counts for sample 2

IncFormLen

Inclusion form length

SkipFormLen

Skipping form length

PValue

P-value for differential splicing

FDR

False discovery rate adjusted p-value

IncLevel1

Inclusion levels for sample 1

IncLevel2

Inclusion levels for sample 2

IncLevelDifference

Difference in inclusion levels between samples (positive = more inclusion in sample 1; negative = more skipping)

Source

Generated with rMATS (https://rnaseq-mats.sourceforge.net/) from K562 eCLIP data.

Examples

data(sample_se.mats)
head(sample_se.mats)
#>   ID          GeneID geneSymbol   chr strand exonStart_0base  exonEnd
#> 1  0 ENSG00000281903  LINC02246 chr21      -        14853273 14853431
#> 2  1 ENSG00000281903  LINC02246 chr21      -        14853273 14853431
#> 3  2 ENSG00000281903  LINC02246 chr21      -        14857541 14857708
#> 4  3 ENSG00000235609       <NA> chr21      -        14857541 14857708
#> 5 14 ENSG00000291050    TEKT4P2 chr21      -         9069153  9069341
#> 6 15 ENSG00000291050    TEKT4P2 chr21      -         9069444  9069599
#>   upstreamES upstreamEE downstreamES downstreamEE ID.1 IJC_SAMPLE_1
#> 1   14825522   14825624     14857541     14857708    0     17,10,13
#> 2   14837181   14837823     14857541     14857708    1      13,8,20
#> 3   14825522   14825624     14902898     14903015    2      13,6,13
#> 4   14825522   14825624     15014343     15014430    3     14,10,15
#> 5    9068358    9068659      9069444      9069599   14        2,0,0
#> 6    9068324    9068659      9070213      9070436   15     23,21,42
#>   SJC_SAMPLE_1 IJC_SAMPLE_2 SJC_SAMPLE_2 IncFormLen SkipFormLen       PValue
#> 1      13,4,13     34,22,21     22,26,29        300         150 0.4074983162
#> 2        0,0,2      32,13,9        6,4,1        300         150 0.0003374221
#> 3        0,0,0     22,26,29        1,0,0        300         150 1.0000000000
#> 4        0,2,0     32,38,37        0,0,2        300         150 1.0000000000
#> 5     12,11,16        0,0,0       4,5,10        300         150 1.0000000000
#> 6      15,13,6      7,13,19        3,5,6        300         150 0.8530224703
#>          FDR         IncLevel1         IncLevel2 IncLevelDifference
#> 1 1.00000000 0.395,0.556,0.333 0.436,0.297,0.266              0.095
#> 2 0.01258623     1.0,1.0,0.833 0.727,0.619,0.818              0.223
#> 3 1.00000000       1.0,1.0,1.0     0.917,1.0,1.0              0.028
#> 4 1.00000000     1.0,0.714,1.0     1.0,1.0,0.902             -0.063
#> 5 1.00000000     0.077,0.0,0.0       0.0,0.0,0.0              0.026
#> 6 1.00000000 0.434,0.447,0.778 0.538,0.565,0.613             -0.019

if (FALSE) { # \dontrun{
  createSplicingMap(bed_file = sample_bed, SEMATS = sample_se.mats)
  createSequenceMap(SEMATS = sample_se.mats, sequence = "CCCC")
} # }